as_seqdict to own file, imported on init
